api/requirements.txt
uvicorn>=0.16.0,<0.18.0
fastapi>=0.64.0,<1.0.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.8.1,<1.10.0
typing_extensions>=3.7.4,<4.0.0
loguru>=0.5.3,<0.6.0
# We will explain this in the course
tid-classifier-model==0.0.2